repopick: Fix repopick with device, kernel, and vendor repos
Currently, the dependencies are set up in the format of:

<project name="DirtyUnicorns/*" remote="github" />

and everything else is of the format:

<project name="*" remote="du" />

which breaks repopick, since Gerrit is setup where everything project
is just its name, not DirtyUnicorns/name.

Before, name would be set to DirtyUnicorns/android_device_huawei_angler
and repopick wouldn't find a proper path as android_device_huawei_angler
is not the same as DirtyUnicorns/android_device_huawei_angler.

After, name is set to android_device_huawei_angler so repopick can find
the proper path as it matches the project tag pulled down from Gerrit.

Change-Id: I1fba0c482f973b7a7f3badc00c9b139db8fe4ea0
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>