config: common: Increase the range of volume steps
We live in an age where media comes in many forms, and often times it is a
challenge to match the correct volume with accuracy due to the limitation
of the number of volume steps provided.

Increase the number of volume steps we currently support by default to 24 in
order to provide a sufficient diversity of volume for media consumption & the
number of in-call steps to 8.

Change-Id: I508f5bcfb83917f7526b6e1eb78e508999812cf1
Signed-off-by: RJ Trujillo <certifiedblyndguy@gmail.com>