roomservice: Make sure we use the right branch
Simple and to the point, we check for external/bson which is
generally associated with our CAF branch to check what branch we're on
and we use that. Nothing fancy but it works.

TODO: Maybe grep our manifest README like we do with other things but
for now, this works.

Change-Id: I3bcd050755bd1a8c8d558eeb2e5833828baf96ee