vendor: strip out packages
Squash of these:

1. Remove LatinIME and 'Basic'

The AOSP keyboard or better known as LatinIME is already set to compile
in the build repo (see link below)

https://github.com/UnicornioSucio/android_build/blob/o8x/target/product/core.mk#L42

As for whatever 'Basic' is, I have no idea. I grep Basic and only thing I found in
a make file was BasicDreams. I don't know what this is building but seems like something
that was in AOSP at one point but no longer needed/wanted.

Change-Id: Ieed755504702cdaab60e1a64685315598838dcab

2. Remove 'development' package

In the past we have compiled this package and just hide it from the app drawer
but really NOBODY uses it. It doesn't serve any purpose these days.

- CellBroadcastReceiver is also removed. Is already being compiled in the build
repo. There's no need for an extra call. Check 'build/make/target/product/telephony.mk'

Change-Id: Ic04044852223a343f2a4b412cd65d0008925a513

3. Most of these tools are already set to compile in the build repo
under the 'core minimal' package which is inherited by the base

https://github.com/DirtyUnicorns/android_build/blob/p9x/target/product/core_base.mk#L64

https://github.com/DirtyUnicorns/android_build/blob/p9x/target/product/core_minimal.mk

OpenVPN is obsolete and we inherited that from the commit below
978aad6

Change-Id: I6145e35587d059fdfe1138436c57fb5cb5110547