Add option to pull pixel gapps conditionally from gitlab
This adds in a very simple but efficient way to include pixel gapps
into devices like the Marlin, Taimen and Nash. Our Gapps package can
not be hosted on Github due to file limits, so this adds in the gapps
package conditionally for those devices.

- Checks for the existence of the pixelgapps repo so we're not attempting
to clone the repo every time.

- Used subprocess to clone repo in order to avoid compatibility issues

- Checks for pixelgapps inheritance line in du.mk to determine whether to
clone the pixel gapps or not. Thanks @DeviousFusion for the idea!

- Add pixel gapps to local manifest so we can pull in updates in a repo sync

Change-Id: I5c2980cf308b563a706b24bd363c10b9946519b8