statix: Implement our accents following the AOSP implementation
This single commit ports all the needed code to support accent colours in Q, it actually is a squash of what we had on Pie with some refinements.
Thanks to POSP and DU for their accent implementation in Pie, the proper commit history can be found in our Pie branch (9.0).

Change-Id: Ic3c71a89f6704d92bb4dbba26d719bfc9a980064