overlay: Allow more cached apps in the background
Google has provisioned this on the Pixel 5 (RQ1A.201205.010) to utilize
the 8 GB of RAM better, and I have also observed that a lightweight
selection of apps was hitting the cached app limit before exhausting RAM
on another device with 8 GB of RAM.

This shouldn't cause adverse effects on devices with less RAM (e.g. 4
GB) as the low memory killer should kick in long before this limit on
such devices.

Change-Id: I97f2aff820b4a4c50657ea63476cc33696027490
Reviewed-on: https://review.statixos.com/c/android_vendor_statix/+/5052
Tested-by: Glitch <humanglitch1.1@gmail.com>
Reviewed-by: Sourajit Karmakar <sourajit@live.com>